removing warnings
git-svn-id: http://code.icecube.wisc.edu/svn/dataclasses/trunk@1371 16731396-06f5-0310-8873-f7f720988828

+++ private/dataclasses/I3Time.cxx
@@ -37,7 +37,7 @@ void I3Time::SetModJulianTime(unsigned int modJulianDay,
 {
   double modjulian = ((double)modJulianDay) + (((double)sec)/(3600. * 24.));
   year_ = yearOf(modjulian);
-  unsigned int daysafteryear = modjulian - modjulianday(year_);
+  unsigned int daysafteryear = (unsigned int)(modjulian - modjulianday(year_));
   unsigned int secsafteryear = daysafteryear * 3600 * 24 + sec;
   daqTime_ = 
     ((long long int)secsafteryear * ((long long int)(1e10)))
